Extract Ultra spatial utils
